+++ BetterFarmAnimalVariety/Framework/Commands/List.cs
@@ -5,7 +5,7 @@ namespace BetterFarmAnimalVariety.Framework.Commands
     class List : Command
     {
         public List(IModHelper helper, IMonitor monitor) 
-            : base("bfav_list", "List the farm animal categories and types.\nUsage: bfav_list", helper, monitor) { }
+            : base("livestock_categories", "List the farm animal categories and types.\nUsage: livestock_categories", helper, monitor) { }
 
         /// <param name="command">The name of the command invoked.</param>
         /// <param name="args">The arguments received by the command. Each word after the command name is a separate argument.</param>